backend/requirements.txt
# Core Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP Client
httpx==0.26.0

# Caching
redis==5.0.1

# Cheminformatics
rdkit==2023.9.4

# Molecular Docking
openbabel-wheel==3.1.1.19  # For PDBQT conversion
meeko==0.5.0  # For ligand preparation
scipy>=1.11.0  # Required by meeko
gemmi>=0.6.0  # Required by meeko

# Task Queue (for async docking jobs)
celery[redis]==5.3.6
flower==2.0.1  # Celery monitoring

# Testing
hypothesis==6.98.3
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Additional Dependencies
python-dotenv==1.0.0
